Fixed a minor issue (Thematic Options will now report the correct status)
Started to code the new feature to speed up Thematic by using constants instead of calling bloginfo() or get_blog_info(). Currently the basics for creating a settings file are integrated. 

I need some feedback from you. Install this version to a live server - I know that it works on a local server, go to Thematic Options and hit 'Create Definitions' .. this will create a directory 'settings' under your child theme's directory and it'll create a file 'settings.php' in this directory. Please report the status .. do you see the directory? Does the file exist? .. or .. failed to create the directory? .. failed to create the file?